+++ inference_perf/client/metricsclient/mock_client.py
@@ -11,8 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from inference_perf.client.modelserver.base import ModelServerMetrics
-from .base import MetricsClient, PerfRuntimeParameters
+from .base import MetricsClient, PerfRuntimeParameters, ModelServerMetrics
 
 
 class MockMetricsClient(MetricsClient):
