+++ inference_perf/client/metricsclient/prometheus_client.py
@@ -14,9 +14,9 @@
 import time
 from typing import cast
 import requests
-from inference_perf.client.modelserver.base import ModelServerMetrics, ModelServerPrometheusMetric
+from inference_perf.client.modelserver.base import ModelServerPrometheusMetric
 from inference_perf.config import PrometheusClientConfig
-from .base import MetricsClient, PerfRuntimeParameters
+from .base import MetricsClient, PerfRuntimeParameters, ModelServerMetrics
 
 PROMETHEUS_SCRAPE_BUFFER_SEC = 5
